+++ app/(home)/_components/WhatsNew/index.tsx
@@ -1,12 +1,43 @@
 import styles from "./index.module.scss";
 import ReactMarkdown from "react-markdown";
 import Block from "@/components/Block";
-import {getWhatsNewRecords, WhatsNewRecord} from "@/lib/whats_new";
 import React from "react";
+import path from "path";
+import fs from "fs";
+import {getSortedPostsData} from "@blog/_lib/load";
+import dayjs from "dayjs";
 
 type Props = {
   id?: string
 }
+
+type WhatsNewRecord = {
+  type: 'page' | 'fix' | 'improve' | 'blog' | 'content',
+  text: string,
+  date: string
+}
+
+const getWhatsNewRecords: () => Promise<WhatsNewRecord[]> = async () => {
+  const jsonPath = path.join(process.cwd(), 'data', 'whats_new.json');
+  const jsonText = fs.readFileSync(jsonPath, 'utf-8');
+  const records = JSON.parse(jsonText) as WhatsNewRecord[];
+
+  const blogData = await getSortedPostsData()
+
+  for (const post of blogData) {
+    records.push({
+      type: 'blog',
+      text: `記事「[${post.title}](https://trpfrog.net/blog/${post.slug})」を公開しました！`,
+      date: dayjs(post.date).format('YYYY-MM-DD')
+    })
+  }
+
+  return records.sort((a, b) => {
+    if (a.date === b.date) return 0
+    return a.date < b.date ? 1 : -1
+  })
+}
+
 
 export default async function WhatsNew ({id}: Props) {
   const whatsNewRecords: WhatsNewRecord[] = await getWhatsNewRecords()
